test case fix
